fix: sign-conversion in array bitmask, restore encode casts for macOS
- art_internal_impl.hpp: cast (i % 8) to unsigned in array-based bitmask
- Restore static_cast<uint64_t> on encode(i) — needed on macOS where
  size_t != uint64_t; suppress GCC useless-cast at file level
- benchmark/micro_benchmark_utils.hpp, micro_benchmark_key_view.cpp:
  same useless-cast suppression

+++ art_internal_impl.hpp
@@ -97,13 +97,17 @@ struct value_bitmask_field<Enabled, std::array<T, N>> {
   std::array<T, N> bits{};
 
   [[nodiscard]] constexpr bool test(std::uint8_t i) const noexcept {
-    return (bits[static_cast<std::size_t>(i) / 8] >> (i % 8)) & 1U;
+    return (bits[static_cast<std::size_t>(i) / 8] >>
+            (static_cast<unsigned>(i) % 8U)) &
+           1U;
   }
   constexpr void set(std::uint8_t i) noexcept {
-    bits[static_cast<std::size_t>(i) / 8] |= static_cast<T>(T{1} << (i % 8));
+    bits[static_cast<std::size_t>(i) / 8] |=
+        static_cast<T>(T{1} << (static_cast<unsigned>(i) % 8U));
   }
   constexpr void clear(std::uint8_t i) noexcept {
-    bits[static_cast<std::size_t>(i) / 8] &= static_cast<T>(~(T{1} << (i % 8)));
+    bits[static_cast<std::size_t>(i) / 8] &=
+        static_cast<T>(~(T{1} << (static_cast<unsigned>(i) % 8U)));
   }
 };
